Distinguish circom tests from JS ones. Add Arnab's string tests
